refactor(scripts): add more utils for deployment and simplify code and config
